backend/requirements.txt
fastapi>=0.136,<1
uvicorn[standard]>=0.44,<1
sqlmodel>=0.0.38,<1
psycopg[binary]>=3.2,<4
pydantic-settings>=2.2,<3
python-multipart>=0.0.25,<1
pwdlib>=0.3,<1
pwdlib[argon2]>=0.3,<1
pyjwt>=2.12,<3